+++ Cargo.toml
@@ -3,7 +3,7 @@
 members = [
     "reductstore",
     "reduct_base",
-    "reduct_macros", "backpack_rs"]
+    "reduct_macros"]
 
 resolver = "2"
