--- backpack_rs/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "backpack-rs"
-version = "0.1.0"
-description = "Backpack is a connector to cloud storage services with caching and local storage capabilities."
-readme = "README.md"
-
-
-authors.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-
-license = "MPL-2.0"
-
-[features]
-default = ["fs", "s3"]
-fs = []
-s3 = ["dep:aws-sdk-s3", "dep:aws-config", "dep:tokio", "dep:aws-credential-types"]
-
-[dependencies]
-url = "2.5.7"
-log = "0.4.27"
-
-# s3 backend
-aws-sdk-s3 = { version = "1.103.0", optional = true, features = ["rustls"] }
-aws-config = { version = "1.8.5", optional = true, features = ["rustls"] }
-aws-credential-types = { version = "1.2.5", optional = true, features = ["hardcoded-credentials"]}
-tokio = { version= "1.47.1" , optional = true, features = ["rt"]}
-tracing-subscriber = "0.3.20"
-
-
-[dev-dependencies]
-rstest = "0.26.1"
