--- backpack_rs/README.md
@@ -1,3 +1,0 @@
-# Backpack
-
-Backpack is a connector to cloud storage services with caching and local storage capabilities.
