--- backpack_rs/src/fs.rs
@@ -1,8 +1,0 @@
-// Copyright 2023 ReductSoftware UG
-// This Source Code Form is subject to the terms of the Mozilla Public
-//    License, v. 2.0. If a copy of the MPL was not distributed with this
-//    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-mod file;
-
-pub use file::{AccessMode, File, OpenOptions};
