+++ reductstore/src/backend/fs.rs
@@ -3,8 +3,8 @@
 //    License, v. 2.0. If a copy of the MPL was not distributed with this
 //    file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use crate::backend::file::AccessMode;
 use crate::backend::StorageBackend;
-use crate::fs::AccessMode;
 use std::path::{Path, PathBuf};
 
 pub(crate) struct FileSystemBackend {
