+++ reductstore/src/cfg/provision/bucket.rs
@@ -1,6 +1,7 @@
 // Copyright 2025 ReductSoftware UG
 // Licensed under the Business Source License 1.1
 
+use crate::backend::BackendType;
 use crate::cfg::Cfg;
 use crate::core::env::{Env, GetEnv};
 use crate::license::parse_license;
@@ -15,7 +16,7 @@ use std::path::PathBuf;
 impl<EnvGetter: GetEnv> Cfg<EnvGetter> {
     pub(in crate::cfg) fn provision_buckets(&self) -> Storage {
         let license = parse_license(self.license_path.clone());
-        let data_path = if self.cs_config.backend_type == backpack_rs::BackendType::Filesystem {
+        let data_path = if self.cs_config.backend_type == BackendType::Filesystem {
             self.data_path.clone()
         } else {
             self.cs_config
