+++ reductstore/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2023-2024 ReductSoftware UG
+// Copyright 2023-2025 ReductSoftware UG
 // Licensed under the Business Source License 1.1
 pub mod api;
 pub mod asset;
@@ -9,3 +9,5 @@ pub mod ext;
 mod license;
 pub mod replication;
 pub mod storage;
+
+pub(crate) mod backend;
